Refactor City Ranking flow
- Update model name from Cidade to City
- Create CityRankingsService to extract logic from the model
